NEW/requirements.txt
# Interfejs użytkownika
streamlit>=1.32.0

# OpenSearch
opensearch-py>=2.4.0

# Embeddingi — stella-pl-retrieval-8k
sentence-transformers>=3.0.0
torch>=2.1.0

# Graf powiązań
networkx>=3.2

# LLM — Groq
groq>=0.5.0

# HTTP (Ollama Cloud)
requests>=2.31.0

# Szablony kontekstu
jinja2>=3.1.0

# Modele danych
pydantic>=2.0.0

# Zmienne środowiskowe
python-dotenv>=1.0.0
